Adjust .profile to use on different OSes
First, I noticed that OpenBSD's shell does not like setting up EDITOR
variable in the shell. Doing so leads to stop working the arrow keys.

Other reason was I put some comments in Polish, and again, OpenBSD
could not display them correctly in the shell. Probably it would've if
I tried xterm and X, for example.